requirements.txt
alembic==1.13.2
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary
python-dotenv==1.0.1
SQLAlchemy==2.0.34
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.1.2
twilio
pyotp